refactor(Validator): move functions to Referee
