--- src/test/java/baseball/ValidatorTest.java
@@ -1,38 +1,0 @@
-package baseball;
-
-import baseball.service.Validator;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-class ValidatorTest {
-
-    @Test
-    void validatePrediction() {
-        // given
-        Validator validator = new Validator();
-        List<String> goodInputs = List.of("123", "456", "789");
-        List<String> badInputs = List.of("", "1234", "112", "12", "012", "zero");
-
-        // then
-        for (String goodInput : goodInputs)
-            assertDoesNotThrow(() -> validator.validatePrediction(goodInput));
-
-        for (String badInput : badInputs)
-            assertThrows(IllegalArgumentException.class, () -> validator.validatePrediction(badInput));
-    }
-
-    @Test
-    void validateReplaySelection() {
-        // given
-        Validator validator = new Validator();
-
-        // then
-        assertDoesNotThrow(() -> validator.validateReplaySelection("1"));
-        assertDoesNotThrow(() -> validator.validateReplaySelection("2"));
-        assertThrows(IllegalArgumentException.class, () -> validator.validateReplaySelection("3"));
-    }
-}
